Add ValidationSelector and RequestAttributeValidator
In order to determine the validation context for different purposes
an abstract implementation is necessary within a neutral attribute
validation environment. This has been added with these new classes.
The RequestAttributeValidator validates a SchemaAttribute in the context
of a clients request and the ValidationSelector selects the type of
validation that has to be executed for the current attribute.