Fix NPE if filtering by meta attribute
SCIM-SDK declares a schema for meta attribute, but it should be
evaluated as part of the main schema according to specification.

Meta attribute was evaluted as extension schema attribute causing NPE in
FilterResourceResolver.retrieveComplexAttributeNode() while filtering
resources.

Mark leaf filter node as main schama node to treat it correctly during
filtering. The mainSchemaNode flag is only used during filtering, so the
change should not break anything.